- Fixed issue with reverse-engineering of index opclasses - Fixed reverse-engineering of partitioning
